Added clear completed function
