+++ src/index.js
@@ -1,8 +1,10 @@
 /* eslint-disable no-use-before-define */
 import './style.css';
+import { removeCompleted, toggleComplete } from './modules/interactive.js';
 
 const newTodoForm = document.querySelector('#new-todo-form');
 const todoList = document.querySelector('.todo-list');
+const clear = document.querySelector('.clearComplete');
 
 function updateTodos(newTodos) {
   const updatedTodos = [];
@@ -31,18 +33,6 @@ const removeTodo = (targetIndex) => {
   const updateList = todoListArr.filter((todo) => todo.id !== parseInt(targetIndex, 10));
   updateTodos(updateList);
 };
-
-function toggleComplete(id) {
-  const todoListArray = JSON.parse(localStorage.getItem('todos') || '[]');
-  const updateTodoList = todoListArray.map((todo) => {
-    if (todo.id === parseInt(id, 10)) {
-      return { ...todo, completed: !todo.completed };
-    }
-    return todo;
-  });
-
-  updateTodos(updateTodoList);
-}
 
 function updateList() {
   const todoListArray = JSON.parse(localStorage.getItem('todos') || '[]');
@@ -89,4 +79,11 @@ function init() {
   updateList();
 }
 
+clear.addEventListener('click', () => {
+  removeCompleted();
+});
+
 init();
+
+// eslint-disable-next-line import/prefer-default-export
+export { updateTodos };
